+++ LeaveManagementSystem/src/app/layout/navbar/navbar.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
 import { Router } from '@angular/router';
 import { AuthenticationService } from 'src/app/core/services/authentication.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-navbar',
@@ -20,9 +21,30 @@ export class NavbarComponent implements OnInit {
   logout(){
     this.authService.logout(this.userId).subscribe((data: any)=>{
       if(data.status ==true){
-        alert(data.message)
-        this.routes.navigateByUrl('/auth/login');
-        localStorage.clear();
+        // alert(data.message)
+
+        const swalWithBootstrapButtons = Swal.mixin({
+          customClass: {
+            confirmButton: "btn btn-success",
+            cancelButton: "btn btn-danger"
+          },
+          buttonsStyling: false
+        });
+        swalWithBootstrapButtons.fire({
+          title: "Are you sure?",
+          text: "You will logout!!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonText: "Yes, signout",
+          cancelButtonText: "No, cancel!",
+          reverseButtons: true
+        }).then((result) => {
+          if (result.isConfirmed) {
+            this.routes.navigateByUrl('/auth/login');
+            localStorage.clear();
+          }
+        });
+
       }else{
         alert(data.message);
       }
